Renaming component sub directory indexs to match component names
Also cleaned up imports and prop deconstruction.
