+++ services/games-spa/src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as Board } from './Board/Board';
+export { default as GameOver } from './GameOver/GameOver';
+export { default as Hint } from './Hint/Hint';
+export { default as HintDialog } from './HintDialog/HintDialog';
+export { default as Word } from './Word/Word';
